Split out all and status into their own WorkflowCount components
