+++ src/lib/components/workflow/workflow-counts.svelte
@@ -5,7 +5,6 @@
   import { workflowStatuses } from '$lib/models/workflow-status';
   import { workflowFilters } from '$lib/stores/filters';
   import { workflowsQuery } from '$lib/stores/workflows';
-  import { workflows } from '$lib/stores/workflows';
   import { decodePayload } from '$lib/utilities/decode-payload';
   import { isStatusFilter } from '$lib/utilities/query/filter-search';
   import { toListWorkflowQueryFromFilters } from '$lib/utilities/query/filter-workflow-query';
@@ -13,9 +12,9 @@
   import { requestFromAPI } from '$lib/utilities/request-from-api';
   import { routeForApi } from '$lib/utilities/route-for-api';
   import { updateQueryParameters } from '$lib/utilities/update-query-parameters';
-
-  import WorkflowCount from './workflow-count.svelte';
-  import WorkflowTypeCount from './workflow-type-count.svelte';
+  
+  import WorkflowCountAll from './workflow-count-all.svelte';
+  import WorkflowCountStatus from './workflow-count-status.svelte';
 
   let totalCount = 0;
   let statusGroups = [];
@@ -50,28 +49,6 @@
       });
     }
   }
-
-  const onTypeClick = (type: string) => {
-    $workflowFilters = [
-      ...$workflowFilters.filter((f) => !isStatusFilter(f.attribute)),
-      {
-        attribute: 'WorkflowType',
-        value: type,
-        operator: 'AND',
-        parenthesis: '',
-        conditional: '=',
-      },
-    ];
-    const searchQuery = toListWorkflowQueryFromFilters(
-      combineFilters($workflowFilters),
-    );
-    updateQueryParameters({
-      url: $page.url,
-      parameter: 'query',
-      value: searchQuery,
-      allowEmpty: true,
-    });
-  };
 
   const onStatusClick = (status: string) => {
     if (status === 'all') {
@@ -149,45 +126,21 @@
   };
 
   $: $workflowsQuery, fetchCounts();
-  $: $workflows, getTopTypes();
-
-  let topTypes = [];
-
-  const getTopTypes = () => {
-    const typeCounts: Record<string, number> = {};
-    const types = $workflows.map((w) => w.name);
-    types.forEach((t) => {
-      typeCounts[t] = (typeCounts[t] || 0) + 1;
-    });
-    topTypes = Object.entries(typeCounts)
-      .sort((a, b) => b[1] - a[1])
-      .slice(0, 5);
-  };
 </script>
 
 {#if groupByEnabled}
   <div class="flex gap-2 lg:gap-4 flex-wrap">
-    <WorkflowCount
-      status="all"
+    <WorkflowCountAll
       count={totalCount}
       {onStatusClick}
       active={!$workflowFilters.length}
     />
     {#each workflowStatuses as status}
-      <WorkflowCount
+      <WorkflowCountStatus
         {status}
         {onStatusClick}
         count={statusGroups.find((group) => group.value === status)?.count ?? 0}
         active={statusFilters.some((filter) => filter.value === status)}
-      />
-    {/each}
-  </div>
-  <div class="flex gap-2 lg:gap-4 flex-wrap">
-    {#each topTypes as [attribute, _count] (`${attribute}`)}
-      <WorkflowTypeCount
-        value={attribute}
-        attribute="WorkflowType"
-        {onTypeClick}
       />
     {/each}
   </div>
